requirement.txt
matplotlib
plotly
langchain-groq
groq
python-dotenv
pandas
streamlit
numpy
joblib
scikit-learn

